@@ -1,8 +1,8 @@
 {
   "name": "ngx-braintree",
-  "version": "0.0.1",
+  "version": "8.0.1",
   "peerDependencies": {
-    "@angular/common": "^6.0.0-rc.0 || ^6.0.0",
-    "@angular/core": "^6.0.0-rc.0 || ^6.0.0"
+    "@angular/common": "^8.0.0 || ^8.0.1",
+    "@angular/core": "^8.0.0 || ^8.0.1"
   }
 }
