@@ -1,5 +1,5 @@
 import { Directive, OnInit, OnDestroy, Renderer2, Inject } from '@angular/core';
-import { DOCUMENT } from '@angular/platform-browser';
+import { DOCUMENT } from '@angular/common';
 
 @Directive({
   selector: '[ngxBraintreeDirective]'
